TCI-582 Fix media_text random empty fields
