+++ config/config-inyo/media.type.snippet.yml
@@ -13,9 +13,7 @@ label: Snippet
 description: 'A reusable block of text.'
 source: text
 queue_thumbnail_downloads: false
-new_revision: false
+new_revision: true
 source_configuration:
   source_field: field_media_text
-field_map:
-  name: name
-  text: field_media_text
+field_map: {  }
